fix: header, regions, news page code
