--- src/views/HomeView.vue
@@ -1,17 +1,0 @@
-<template>
-  <div class="home">
-    <main-page />
-  </div>
-</template>
-
-<script lang="ts">
-import { defineComponent } from 'vue'
-import MainPage from '@/components/Main.vue'
-
-export default defineComponent({
-  name: 'HomeView',
-  components: {
-    MainPage
-  }
-})
-</script>
