+++ tsconfig.json
@@ -5,9 +5,9 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["./src/*"]
-    }
+    },
+    "lib": ["ES2017"]
   },
-
   "references": [
     {
       "path": "./tsconfig.node.json"
